dom/base/test/test_meta_viewport5.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>meta viewport test</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <meta name="viewport" content="user-scalable=NO">
  <script src="viewport_helpers.js"></script>
</head>
<body>
  <p>user-scalable=NO</p>
  <script type="application/javascript;version=1.7">
    "use strict";

    SimpleTest.waitForExplicitFinish();

    let tests = [];

    tests.push(function test1() {
      let info = getViewportInfo(800, 480);
      is(info.allowZoom, true, "user-scalable values are case-sensitive; 'NO' is not valid");

      nextTest();
    });

    function getViewportInfo(aDisplayWidth, aDisplayHeight) {
      let defaultZoom = {}, allowZoom = {}, minZoom = {}, maxZoom = {},
          width = {}, height = {}, autoSize = {};

      let cwu = SpecialPowers.getDOMWindowUtils(window);
      cwu.getViewportInfo(aDisplayWidth, aDisplayHeight, defaultZoom, allowZoom,
                          minZoom, maxZoom, width, height, autoSize);
      return {
        defaultZoom: defaultZoom.value,
        minZoom: minZoom.value,
        maxZoom: maxZoom.value,
        width: width.value,
        height: height.value,
        autoSize: autoSize.value,
        allowZoom: allowZoom.value
      };
    }

    function nextTest() {
      if (tests.length)
        (tests.shift())();
      else
        SimpleTest.finish();
    }
    addEventListener("load", nextTest);
  </script>
</body>
</html>
